Rag Final Project/rag_customer_support/requirements.txt
streamlit>=1.34.0
python-dotenv>=1.0.1
pypdf>=4.2.0
langchain>=0.3.0
langchain-community>=0.3.0
langchain-openai>=0.2.0
langchain-huggingface>=0.1.0
langchain-chroma>=0.1.2
langchain-text-splitters>=0.3.0
langgraph>=0.2.0
sentence-transformers>=2.7.0
typing-extensions>=4.9.0


